reduced initialization time and fixed a 32-bit timestamp I missed earlier.  May have fixed #96 but it's hard to tell

+++ src/mavlink_stream.c
@@ -19,7 +19,7 @@
 typedef struct
 {
   uint32_t period_us;
-  uint32_t last_time_us;
+  uint64_t last_time_us;
   void (*send_function)(void);
 } mavlink_stream_t;
